Enemy 2 Animations, collisions, better ground finding and rotation
The enemy is now almost complete, all he needs now is health and attacking (which shouldn't be too hard), then we can start using him in the map for real!
Slight bug with randomly jumping, which i will try to solve, but its not a huge deal at the end of the day.